requirements.txt
numpy
matplotlib
Pillow
keras
wxPython
